Bigint in most.

+++ db/migrate/20190611094341_create_medication_prescriptions.rb
@@ -1,7 +1,7 @@
 class CreateMedicationPrescriptions < ActiveRecord::Migration[5.2]
   def change
     create_table  :medication_prescriptions, :primary_key => :medication_prescription_id do |t|
-    	t.integer    :drug_id
+    	t.bigint    :drug_id
       t.bigint      :encounter_id
     	t.datetime   :start_date
     	t.datetime   :end_name
