+++ db/migrate/20190611095055_create_patient_histories.rb
@@ -2,8 +2,8 @@ class CreatePatientHistories < ActiveRecord::Migration[5.2]
   def change
     create_table    :patient_histories, :primary_key => :history_id do |t|
     	t.bigint       :encounter_id
-    	t.integer       :concept_id
-    	t.integer       :value_coded
+    	t.bigint       :concept_id
+    	t.bigint       :value_coded
     	t.boolean     :voided, null: false, default: 0
     	t.bigint        :voided_by
     	t.datetime    :voided_date
